Pool2D implementation for CPU backend with mkldnn
Summary:
Implements pooling with the CPU backend with MKL-DNN. Supports max pooling and average pooling (with and without padding).

Unless the forward pass is done in inference mode, mkldnn requires that both workspace memory and a pooling primitive descriptor used in the forward pass be passed to backwards primitive descriptors so as to better optimize gradient computation. Memory descriptors for both of these are captured/saved on the stack for gradFunc, since they can't be reconstructed without the forward descriptor.

Reviewed By: vineelpratap

Differential Revision: D13579084

fbshipit-source-id: 4a7521644f11352f710b2d27eeddd90127e93d95