trainable frontend v2
Summary:
Finally got bandwidth to get back to trainable frontend. This is just a refactor of D12995638 which is compatible with latest version.

I will add the code to generate gabor/ gammatone filters on the fly as a stacked diff next, so we are not restricted to 80 filters.

Then we can add support for arch file so we can finally train the models.

Reviewed By: vineelpratap, jacobkahn

Differential Revision: D13854330

fbshipit-source-id: bd84a5152b63f4df9154dc17472e53203429b042