blacklist cuDNN conv algo 1
Summary:
Certain Librispeech runs were failing on P100 GPUs but not on V100. I traced the issue back to cuDNN. In cuDNN backward, one is allowed to select an algorithm. When using `CUDNN_CONVOLUTION_BWD_DATA_ALGO_1`, the call to `cudnnConvolutionBackwardData` can randomly produce incorrect results with huge discrepancies.

It appears to be a Heisenbug, dependent on the context of the call. Running the `cudnnConvolutionBackwardData` with the same inputs in isolation may not reproduce the error. I even tested calling it in a loop, and usually it succeeds on the 2nd attempt (with the exact same inputs!).

This diff's fix is to blacklist `CUDNN_CONVOLUTION_BWD_DATA_ALGO_1` and also `CUDNN_CONVOLUTION_BWD_FILTER_ALGO_1` for safety. Test plan shows this works and doesn't incur a significant performance hit.

Reviewed By: vineelpratap

Differential Revision: D14960116

fbshipit-source-id: 89ca2153ce4ff6dcb12f11a001d8b41f82990c9e