Add coalescing reducer and async cached gradient sync
Summary:
Implements a `CoalescingReducer` in flashlight distributed and sets it as the default reducer in wav2letter for gradients. Adds testing and use of the reducer in an example.

Changes `allReduceSet` such that a set of passed tensorsmust be of the same type if using contiguous mode. Since only one `allReduce` call is made, only one `ncclDataType_t` can be used for the entire buffer.

Reviewed By: andresy

Differential Revision: D14819005

fbshipit-source-id: 56b83552ec4bf03fef4d3efa908484ebf25c59fe