Force rtd to use requirements for sphinx
Summary:
The docs build is broken yet again because of some change to pip cache on rtd containing a version of Sphinx that isn't compatible with what Breathe is generating. Our docs build is now 2+ weeks stale. Rather than relying on the cache, add this config to [hopefully] force rtd to do a clean install of the needed dependencies every time it builds the docs.

(Note: this ignores all push blocking failures!)

Reviewed By: xuqiantong

Differential Revision: D15509742

fbshipit-source-id: e82c437e99c828aac121ab6b5a1c77139edcd717