Make DevicePtr movable, use in flashlight distributed
Summary:
Make `DevicePtr` movable so that it can be used in some limited data structures and passed around. Since we don't want it to be copyable, we have to use an `unordered_map` rather than a `vector` to store mappings for array sizes in `allreduceMultiple`.

Switch to using `DevicePtr` for single and multiple array cases for CUDA and CPU backends.

Reviewed By: jcai1, vineelpratap

Differential Revision: D15513654

fbshipit-source-id: 99d04cb7def1f97e3b9cd1441f76f377512ff075