CircleCI automated build checks and CUDA/CPU Dockerfile base
Summary:
Add CircleCI integration for the CUDA and CPU backend builds. We need this: some recent w2l PRs have broken the build, and incentives unfortunately aren't in place to import a PR, patch the diff, and run tests ourselves. Until all of our tests run on internal CI resources (work in progress, almost there), this will sanity check that the PR builds and give a signal to open source contributors as to whether or not their changes build.

If this works well and contributions increase, we'll add dedicated AWS GPU resources to run flashlight tests direclty through CircleCI; this is a first step.

Core changes to Dockerfiles:
- Update ArrayFire verison to v3.6.4, add recursive submodule update
- Take advantage of Docker caching and create more intermediate steps for recovering more quickly on failed builds
- Create `Dockerfile-CUDA-Base` and `Dockerfile-CPU-Base`
    - This facilitates creating an environment in which we can test flashlight from scratch without pre-building. We want CircleCI to do the build and succeed/fail, so stop building directly in the Dockerfile, else CircleCI will output a generic error with the Dockerfile environment failing to initialize
    - `Dockerfile-CUDA` now inherits directly from `Dockerfile-CUDA-Base`, same with CPU

In order for the CircleCI build to run properly, Docker Hub assets need to be updated, which I'll do concurrently with landing this diff. Will be added to wav2letter in a separate diff pending these working

Reviewed By: tlikhomanenko

Differential Revision: D15512063

fbshipit-source-id: b54826f97aed7efeed4733c128df76d583fea451