add composite memory allocator
Summary:
Create composite allocator with sufficient memory for all allocation but insufficient memory in smallest buckets such that small allocation are propagated to larger buckets.
Allocate memory in exponential distribution of size and random order of deallocations. Verify that allocation are within the arena and that allocator accounts for allocated and freed memory.

Reviewed By: padentomasello

Differential Revision: D19964152

fbshipit-source-id: ede7579af9055e1b24f2574e878521d5d11b9f54