Fix undefined behavior in grad clipping test
Summary:
Somehow, passing raw doubles to `fl::allClose(const af::array& a, const af::array&b, ...)` compiled, ran, and usually produced the correct result.

This is actually undefined behavior - when I do this again in another test, it doesn't work.

Unclear how this happened, but you get a prize if you can figure out how the compiler thought it was ok to cast a double to a `const af::array&` - it isn't even clear it envoked a default constructor

Reviewed By: vineelpratap

Differential Revision: D21296685

fbshipit-source-id: 045e1ff2ff8cd5b8408d0360568aacea4ae7778e