Enhance the CachingMemoryManager to possibly handle high memory pressure (#188)
Summary:
Enhance the CachingMemoryManager to handle high memory pressure by adding 2 options.
Note: the default behavior is unchanged.
Add a new API to MemoryManagerAdapter to setOption().
Side: remove some glog includes failing compilation if glog-dev not installed locally

**Original Issue**:
https://github.com/facebookresearch/flashlight/issues/180
closes https://github.com/facebookresearch/flashlight/issues/180

The CachingMemoryManager being greedy in memory. This PR adds 2 options to mitigate this.

### Test Plan (required)
Implemented a new unitest in order to show OOM with the default CachingMM and no OOM if the right option is set.

Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/188

Reviewed By: vineelpratap

Differential Revision: D24435873

Pulled By: jacobkahn

fbshipit-source-id: 99186439c1e306ad987079cc326ab68fe1f028fc