Enforce type consistency across non-unary autograd operators (#116)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/116

Because of changes made in D21235273 that require non-casting operators in the computation graph to share gradient dtypes, `backward()` might throw if computing a gradient with respect to an operator that took two `Variable`s of different types.

This behavior is weird, so we want to throw on forward and alert the user that they are performing incompatible operations - in particular, operations, that can't be differentiated through.

There are two scenarios in which the user can do this properly:
- Use `Variable::as` as introduced in D21235273 to add another node to the computation graph
- Manually cast `af::array`s themselves and leave upcasting/downcasting to ArrayFire, e.g.:

```
auto f16 = Variable(af::randu({5, 5}, af::dtype::f16), true);
auto f32 = Variable(af::randu({5, 5}, af::dtype::f32), true);

auto outWithMixedTypeAdding = f16.array() + f32.array();
```
Note that the above doesn't create a differentiable structure since the output Variable isn't connected to the input Variables in the computation graph. `Variable::as` enables this

Reviewed By: vineelpratap

Differential Revision: D21360292

fbshipit-source-id: 2c6abd91a2497f5eb2a1d66b74aae91631cf28c0