Fix header install paths, globbing, exclusion, and config params (#233)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/233

Clean up which headers we install.

We use `install(DIRECTORY ...)` with `FILES_MATCHING`, which is the easiest way to recursively move headers in CMake.

The problem is (yep, this is real) that empty directories that don't have matching patterns are still installed. [There's a ticket for this](https://gitlab.kitware.com/cmake/cmake/-/issues/17122) but lord knows when it'll be merged in. Either way, we need to support old behaviors.

Adds exclusive patterns for things we don't want to install to avoid having empty directories (vcpkg also prohibits having empty directories).

This patterns are a little scary (reasonable so far) -- if we need to add more things that don't fit with these patterns well, we can modify the vcpkg port to manually remove the directory.

The only alternative here is to use CMake's `file(GLOB_RECURSE ...)` to manually locate headers, but this doesn't preserve relative paths and creates a ton of headaches.

Reviewed By: tlikhomanenko

Differential Revision: D24720661

fbshipit-source-id: c73ef1b6fba653fded4d175572a50b0189d561db