Use private source properties rather than build interface gen exps (#264)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/264

**tl;dr** - building with 8 threads is 4x faster and only 1.7x slower than building with 40/80 threads.

Don't use generator expressions because they get expanded per-target which is slow and means that source files can't be re-used in object form across targets with the same dependencies.

Make everything private instead. This used to be a problem when we used interface libraries because those libraries weren't in the export sets of installed targets which meant paths prefixed with the source directories couldn't be exported. This doesn't apply anymore given that we have single libs.

Reviewed By: tlikhomanenko

Differential Revision: D25008791

fbshipit-source-id: c10c71de25bb80489326ea068d944605845baffa