Fix FindGFLAGS to match case
Summary:
New versions of CMake such as those used with `vcpkg` (>= 3.18) treat `find_package` modules differently when the case of the name of the module doesn't match the argument passed to `find_package_handle_standard_args` that sometimes leads to failure. Update the casing of `gflags` (not `GFLAGS`) to match the library names.

`Findgflags.cmake` is added back in D25092173 since case-sensitive file renames are properly tracked by mercurial

Reviewed By: tlikhomanenko

Differential Revision: D25092161

fbshipit-source-id: 16b7156522bdb74f438028ba6dd7118ba1b136ea