fix dataset bug in case of empty target transcription in the list file (#279)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/279

i) fix creation of target if it is empty (no af exception on empty data)
ii) fix join operation (for batching) when the whole batch has empty transcription - in this case just create the whole pad matrix

We need to have this solution because depending on the mapping with lexicon some targets can become empty and filtering this out is not comfortable with list filtering. Either we add in future their filtering or have solution with putting pad token for such samples and processing pad token by loss in a right way.

Reviewed By: vineelpratap

Differential Revision: D25157814

fbshipit-source-id: f697e5468beb1edbfad4dc984e6d7ea86933f206