Add experimental advanced indexing operator in flashlight (#27)
Summary:
This PR adds an advanced indexing operator that is able to provide the correct gradient when indices are repeated in the input to the operator. The operator is limited in the sense that is only able to apply advanced indexing to the second dimension for a 2-dimension tensor.

Pull Request resolved: https://github.com/fairinternal/flashlight/pull/27

Reviewed By: tlikhomanenko

Differential Revision: D24936310

Pulled By: chaitan3

fbshipit-source-id: 7301079e346dbdd17faa320d1dfe46dbc5bf643b