Fix Docker/CI issues with test discovery (#317)
Summary:
CMake's gtest_discover_tests actually runs the test and asks for a list
of tests. In some cases, when special initialization happens in the test
(as in the AllReduce or CachingMemoryManager tests), this fails because
either `mpirun` isn't in the path, libfabric is missing, or there's no
CUDA driver on the machine doing the build.

These shouldn't be requirements of building, so make test discoverability
configurable when adding the test - don't do this for the two aforementioned
tests.

Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/317

Reviewed By: tlikhomanenko

Differential Revision: D25432059

Pulled By: jacobkahn

fbshipit-source-id: 0c48a399c1153c2538e4eb3d2521faafc593d26d