Add fp16 type, fix conv precision in tests (#319)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/319

Making sure autograd tests fail properly — need to pass DNNL an fp16 type for it to throw properly, otherwise we opaquely pass an fp16 buffer as an fp32 type which segfaults

TODO: add some sort of `fl::f16Supported()` which wraps both ArrayFire and whether or not we're using the CPU backend (add a macro)? Will do this in a future change.

Reviewed By: tlikhomanenko

Differential Revision: D25455157

fbshipit-source-id: 21ec7b3c2a579fdad25dc7f1e0b8da3722b5e030