RNN, LSTM, LBR GRU oneDNN implementations matching cuDNN
Summary:
RNN inference on the CPU backend. Lets us do inference with seq2seq models.

The huge challenge here was figuring out how to use cuDNN weights for oneDNN inference in such a way that the results were the same. To my knowledge, no other framework has gotten this interoperability to work. This was a disaster because the cuDNN weights tensor is 1D with an undocumented order and layout, and it has to be converted to several different oneDNN tensors, some of which are 5D, with 5D reorder operations. cuDNN also has two biases for RNN cells by default whereas oneDNN only has one, so those need to be combined. Gradient computation, when implemented later (if we have a need for seq2seq finetuning) will be easy, as we can distribute half of the total bias gradient computed by oneDNN to each of the cuDNN gradients.

I'm going to start enabling CPU tests and properly gating unsupported features to be skipped. Since we're using the CPU backend for more inference-related things, having tests actively running for it will be better support/that we pay attention. The community continues to use it quite a lot.

My attempt at something that properly does bidirectional weights is here: https://github.com/jacobkahn/flashlight/tree/bidirectional_attempt -- the complexity of converting cuDNN weights to oneDNN weights too large for now. I may return to this later.

I plan on also adding OpenCL interoperability for AMD GPUs on top of oneDNN at some point - this will probably come later since there are some other decisions with how we build variosu backends depending on which hardware the user has, and those things should probably be made configurable.

Differential Revision: D25561730

fbshipit-source-id: 7746013e27dad4ed705892d056678d2d1e905cbe