Colab dependency stub script
Summary:
Provide a stub script that sets up the environment and dependencies in a colab env. This will serve as a single source of truth and can be updated as we go. This isn't a general script - this really only applies to the weird Ubuntu 17.04 colab env with CUDA 10.0.

NB: env vars at the bottom need to be set for flashlight before building.

I'm not including FL-specific build commands here because they might change based on what the user wants to build, and the user should select backend, apps, etc.

Reviewed By: tlikhomanenko

Differential Revision: D25684276

fbshipit-source-id: 01f751fa5d59b1bd9b0c8039010b154a339ab1f3