Add fl::init(), remove memory manager global static initialization
Summary:
Due to some CUDA 11 changes in how device context initialization occurs, `af::DeviceManager` initialization behaves differently so as to create a race with the global static initialization of the MemoryManagerInstaller setter.

Remove the global initialization to remove the race and initialize manually in binaries. Add `fl::init()` and call it everywhere.

Reviewed By: vineelpratap, tlikhomanenko

Differential Revision: D25805375

fbshipit-source-id: ba7a08a0fe6eab27da7e5200869d452e5e427015