Reduce recycling size limit and split size limit of the CachingMemoryManager
Change CachingMemoryManager’s size of recyclingSizeLimit_ and splitSizeLimit_  from infinity
to 256MB and 512MB respectively. These values fixed OOM crash on testing of dynamic batching.
These values are reasonable for a lot or all of our workloads based on memory allocation analysis
we run over a number of models. OOMs that we analysed in the past failed on allocation of large
buffers (~ 50MB..450MB). Often this memory was available but fragmented.