fix the stability of multiheadAttention for fp16 computation
Summary: For fp16 training at some point (even to at the end of training) loss NaN occurs, which is caused by -inf values for all axis for some Time position in self-attention (after softmax this will give nan for attn weight). This is caused by the overflow of matmulNT(q, k) and the same for rel pos emb, which produce these -inf. The fix is to performs scaling before linear operations.

Differential Revision: D26177968

fbshipit-source-id: 96ddd40b94ca6b155a53bced9d70ab5be1432b0f