Add Findsox.cmake and Buildsox.cmake (#471)
Summary:
Pull Request resolved: https://github.com/facebookresearch/flashlight/pull/471
If we have libsox on the machine(s) that builds the and run FL, we should use Findsox to find and link with the libsox shared library.
Otherwise, we should Buildsox and link statically.
If we want to link with preinstalled libsox static library, we will have to link with a number of other libraries. So, to keep things simple, we download and build it instead.
Buildsox source code is based on PyTorch cmake script at:
https://github.com/pytorch/audio/blob/2c8aad97fc8d7647ee8b2df2de9312cce0355ef6/third_party/sox/CMakeLists.txt

Differential Revision: D26595938

fbshipit-source-id: 1f9dc8099c7286c6d1916e50dde9a008badc9647