Add sum reduction -- fix default reduction behavior to match numpy and add multiple dimension reductions.
Summary:
This implements the sum reduction, modifies amax and amin to work across multiple dimensions, and fixes the behavior of them to match numpy if no axes are given.

After playing around with std::optional for a bit, I realized it may not support the API we'd like. You can't use std::optional with const references, so doing

sum(arr, { 0, 1}) fails.

Please correct me if I'm wrong or missing something.

Ultimately, I opted for function overloading instead which gives us a nice API at the expense of a few more function definitions. (Which I think we could solve with the preprocessor similar to how you did binary ops if need be but it's really somewhat trivial).

Reviewed By: jacobkahn

Differential Revision: D28395739

fbshipit-source-id: 22a8d4cdf3d90a7364f196c7948332079c5f4fe8