Add Shape to ArrayFireTensor, remove other shape implementations (#606)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/606

Pull Request resolved: https://github.com/fairinternal/flashlight/pull/39

Introduce a shape implementation that isn't dependent on ArrayFire and can be used as an implementation stub for future tensor implementations.

This removes `Shape` as a wrapper for `af::dim4`, and instead leans on explicit conversions from `af::dim4` to `fl::Shape`.

`fl::Shape` *doesn't follow the same shape conventions as does ArrayFire* -- the shapes `{3, 1}` and `{3}` are *not* equivalent. There are no dimension one defaults, and ArrayFire dimensions are converted accordingly (see `ShapeTest` for some examples).

Adds `Shape` as a member on `ArrayFireTensor` — now, `Tensor::shape()` can return a const ref.

Reviewed By: benoitsteiner

Differential Revision: D28547964

fbshipit-source-id: e73e138b690c06957a86bfa8a7fe7fb31d14ade4