Introduce TensorBackend (#622)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/622

Adds a `TensorBackend` abstraction. This allows the following:
- Global functions that operate on tensors (and aren't members of `fl::Tensor`) need to call into corresponding implementations based on the backend of the tensor operands.
- Tensor backends need to standardize their implementations and implement certain operators. Enforcing this with missing symbols in a global function API isn't really feasible; using pure virtual functions accomplishes this more easily
- Facilitates storing global state associated with a particular tensor backend (e.g. a )

The dispatch pattern for global tensor operations is as follows:
1. A global function operating on a tensor is called, e.g. `fl::log(tensor)`.
2. `fl::Tensor::backend()` is used to ascertain the `TensorBackend` instance for that given tensor based on its `TensorBackendType`
3. the `fl::TensorBackend::log()` function is dispatched to for the tensor's backing `TensorBackend`.

While this adds an additional layer of runtime polymorphism, the overhead is minimal; a single `switch` is required to find the `TensorBackend` per tensor (branch prediction should be solid for 99% of cases) and one virtual function call is required to dispatch to the correct function implementation.

`s/TensorBackend/TensorBackendType/g`

I've moved `log` and `exp` as sample operators so far; will move other operators in a future change — this is to get input on the design.

Differential Revision: D28659887

fbshipit-source-id: dd7510b431513d7f37f760bb0d7b1c1fb67a3698