Assignment operators, generics reorg
Summary:
Add assignment operators for tensors and in-place arithmetic operations.

These are implemented as functions (e.g. `inPlaceAdd`) on `TensorBackend` and operators (`operator+=`) on `Tensor` for clarity and to make the `TensorBackend` API easier to derive from.

Also move templated reductions from `ArrayFireTensor` into `TensorBase` and manually define specializations in `TensorBase`. For now, these specializations simply cast the result of the reduction defined on `TensorBase` to the relevant type — the functions on `TensorBase` return `double` as needed.
- Eventually, we should remove/reimplement these specializations in terms of the reductions that return tensors then let the user decide how to cast the tensor into some type on host memory.

Also define a generic version of `allClose` that operates on `fl::Tensor`/use it in tests.

Differential Revision: D29118744

fbshipit-source-id: 157e8ba4cb45c45158867b491b0f2c6bce6e7ab1