Fix end idx obo
Summary:
`fl::end` corresponded to index `-1` but was being translated to `-2` via the `range` constructor. To distinguish between `end` and other indices, add another type `end_t` which decays to -1 and have range ctors take a variant of the two.

`end_t` naturally decays to -1 via the matched `int` coercion in `visit`.

Differential Revision: D29318965

fbshipit-source-id: d68e6ffbd7316d48769a7cdc8263ce89949fe11c