Fix end idx obo (#647)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/647

`fl::end` corresponded to index `-1` but was being translated to `-2` via the `range` constructor. To distinguish between `end` and other indices, add another type `end_t` which decays to -1 and have range ctors take a variant of the two.

`end_t` naturally decays to -1 via the matched `int` coercion in `visit`.

Reviewed By: benoitsteiner

Differential Revision: D29318965

fbshipit-source-id: ed41dd45aea7b1fe6a12f43a26ccc77a04572347