Add proxy views and fix lvalue indexing for ArrayFireTensor (#649)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/649

`ArrayFireTensor::index()` internally returned an `af::array::array_proxy` when `af::array::operator()` was used, but that was auto upcast to an `af::array` via `af::array::array_proxy::operator array()`, which caused a copy on write semantic that precluded proper in-place mutations for lvalue `fl::Tensor`s.

The fix is to allow `ArrayFireTensor` to also store an `af::array::array_proxy` as might be needed for some lvalues. As soon as an `af::array` handle is requested or some other op that requires an `af::array` is called, `ArrayFireTensor::handle()` upcasts the proxy to an `af::array` via the aforementioned operator.

This is implemented via a `variant` storing both array and proxy types, and methods that are on either are called as visitors.

Because getting an array handle now might mutate the `ArrayFireTensor` internals via promotion from a proxy to an array, some constness fixes were needed throughout. Methods on `fl::Tensor` remain const, but methods on `TensorAdapter` don't have const requirements anymore.

Reviewed By: benoitsteiner

Differential Revision: D29356855

fbshipit-source-id: 4bd86cdc714bfd6c702a475c940f07d4e4ba922d