Rewrite af-based indexing (#657)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/657

Turns out storing an `af::array::array_proxy` as an lvalue isn't really supported and creates lots of problems. Instead, create a visitor which returns an array proxy as a temporary lvalue if needed for in-place operators (assignment, etc) that also returns an array ref.

This is essentially deferred, lazy indexing where indexing occurs after the tensor is used as an rvalue. lvalue tensors that hold proxies can still be operated on as in the case of-in-place assignment.

Also add some explicitly-defined move ctors to avoid some implicit copies that were happening in `Index` and `Tensor`.

Reviewed By: padentomasello

Differential Revision: D29420741

fbshipit-source-id: ebfb2f5e1258afa0a1028957ae4f8d654201bc5d