Conformer: (breaking change) remove redundant skip connections and fix pointwise conv (#587)
Summary:
**Original Issue**: https://github.com/flashlight/flashlight/issues/582

Please see the linked issue. This PR fixes redundant skip connections in Conformer, as a breaking change. Early tests show this may impact model accuracy by 10% (relative) or more.

This PR also fixes a bug where the output of the first pointwise convolution in a Conformer conv block was unused. The accuracy impact of fixing that bug is not yet known.

I believe as part of merging this PR, the RASR readme should be updated to link to the last flashlight commit where the old RASR Conformer models are expected to work, and new RASR conformer models should be trained and released with the fix.

_The alternative to a breaking change here would be to come up with some kind of constructor hack for the Conformer block to make old models load while applying the fix to new models. That wouldn't be worth it for me though. I'd rather retrain my models to get the improved model accuracy - that's the whole reason I was using a Conformer in the first place._

Pull Request resolved: https://github.com/flashlight/flashlight/pull/587

Reviewed By: tlikhomanenko

Differential Revision: D29891724

Pulled By: jacobkahn

fbshipit-source-id: 1a401502d89813918c9ff9dc7b573eef54e36a3b