Tensor backend conversion - toTensorType (#691)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/691

Backend conversion across Tensors.

The implementation requirement for `fl::Tensor::device<T>()` stipulates that memory isn't freed if the underlying tensor falls out of scope unless `unlock` is called — means a device pointer can be used even after the tensor dies since the func takes a `Tensor&&`.

UB can occur if the user creates a shallow tensor copy then converts one of the shallow copy into another tensor (depending on how each backend manipulates the underlying memory), but this could be useful in some cases to save memory, etc, so left as is for the user to control.

Reviewed By: benoitsteiner

Differential Revision: D29856404

fbshipit-source-id: 8062216eebc798c283f697c0dbf8bd1ed5181081