Add runtime switching of tensor backends (#692)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/692

See title.

Since tensor types are parameterized by derived types of `TensorAdapterBase`, the argument to `setDefaultTensorBackend<T>()` is a type. The type gets captured and stored via a closure on `DefaultTensorType` which returns a pointer to the adapter created with the requisite params.

Usage (see test):
```
fl::setDefaultTensorType<ArrayFireTensor>();
// ...
fl::setDefaultTensorType<FooTensor>();
```

Reviewed By: benoitsteiner

Differential Revision: D29869330

fbshipit-source-id: edf573557712d79003e19999cbe6ec4b4d98adb5