Fix gradient computation for batched matrix multiply (#760)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/760

Gradients for batched matmul are of incorrect shapes — fixes the issue. What's happening here:

Consider `matmul(lhs, rhs)` where `lhs = Shape(10, 12)` and `rhs = Shape(12, 14, 2)`. The result is of shape `(10, 14, 2)`:

wlog, the gradient of the lhs is `matmul(gradOutput, rhs^T)` (`^T` --> transpose) which is `(10, 14, 2) x (12, 14, 2)^T --> (10, 14, 2) x (14, 12, 2) --> (10, 12, 2)`

Before, we were adding a gradient of size `(10, 12, 2)` to a tensor of size `(10, 12)`, which is incorrect. Since batched matmul simulates a tile operation in the batch dimension(s), the gradient for a non-batched input should be the sum over the tiled batch dimensions --> `sum((10, 12, 2), dim=2)` gives the correct output

Differential Revision: D31966446

fbshipit-source-id: a772a1a6cdc30a0a460c73cb92322823cf699ed6