Add AutogradExtension for autograd tensor ops
Summary:
Adds an `AutogradExtension` based on `TensorExtension`. Implements operations commonly found in autograd-aware deep learning packages (i.e. cuDNN, oneDNN, miOpen, SYCL-standard things, etc) under a set of common interfaces that can be conditionally compiled.

In particular, this also exposes autograd ops to the common Flashlight Tensor interface, such that ops like `conv2d` can be used on tensors which are not autograd-aware (are not `Variable`s).

The ops that are not integrable into this interface include:
- conv2d
- pool2d
- batchnorm
- rnn

(and `backward` variances thereof, computing gradients with respect to various inputs)

For now, create a `CudnnAutogradExtension` — will add a `OnednnAutogradExtension` in a later change.

Differential Revision: D32611515

fbshipit-source-id: 456b01f690e2ae06c0cb837ec6dac29f79ead899