Fix empty target export list, remove core dep on lib (#834)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/834

Resolves https://github.com/flashlight/flashlight/issues/814 --- uses a proper CMake list collection and appends targets in all cases.

Also removes the dependence of Flashlight on `fl_lib_common`, which means we can clean up and move these things out of the Flashlight core as needed.

Introduces use of `std::filesystem` (which we already have since FL is C++17 minimum outside of CUDA kernels which are C++14 until we get a new nvcc version that supports GCC 9 which requires using CUDA 11 minimum) which removes the dep on `fl::lib` file blah in `lib/common/System.h`.

Differential Revision: D34811636

fbshipit-source-id: 0f1dfebc5a52b855ca71fed4395701ffaa796036