Add AutogradExtension for autograd tensor ops (#778)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/778

Adds an `AutogradExtension` based on `TensorExtension`. Implements operations commonly found in autograd-aware deep learning packages (i.e. cuDNN, oneDNN, miOpen, SYCL-standard things, etc) under a set of common interfaces that can be conditionally compiled.

In particular, this also exposes autograd ops to the common Flashlight Tensor interface, such that ops like `conv2d` can be used on tensors which are not autograd-aware (are not `Variable`s).

The ops that are not integrable into this interface include:
- conv2d
- pool2d
- batchnorm
- rnn

(and `backward` variances thereof, computing gradients with respect to various inputs)

For now, create a `CudnnAutogradExtension` — will add a `OnednnAutogradExtension` in a later change.

Differential Revision: D32611515

fbshipit-source-id: d4da6bfd78e7300f83a3a7b8e027bf1789a1a22d