Allow callback registration for Device::setActive (#906)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/906

As title, `Device::addSetActiveCallback` allows us to move the ArrayFire-active-device-context-synchronization logic into a callback in `ArrayFireBackend` (which will subsequently be added onto relevant devices.

This enables us to remove the `runtime(CUDADevice) --> tensor(ArrayFireBackend)` dependency.

Next step is to completely remove `runtime --> tensor` dependency, in order to prevent a dependency cycle that would've been introduced in an upcoming diff.

Differential Revision: D37465389

fbshipit-source-id: 618d91036c50d2b21c6a455e58e37120d0a3fa29