Allow callback registration for Device::setActive (#906)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/906

As title, `Device::addSetActiveCallback` allows us to move the ArrayFire-active-device-context-synchronization logic into a callback in `ArrayFireBackend` (which will subsequently be added onto relevant devices.

This enables us to remove the `runtime(CUDADevice) --> tensor(ArrayFireBackend)` dependency.

Next step is to completely remove `runtime --> tensor` dependency, in order to prevent a dependency cycle that would've been introduced in an upcoming diff.

Reviewed By: jacobkahn

Differential Revision: D37465389

fbshipit-source-id: d47742f85b7ce4ca8c898f2e39f0d7021dab4894