Use a newly created stream for cudnn via the runtime abstractions (#920)
Summary:
Pull Request resolved: https://github.com/flashlight/flashlight/pull/920

Instead of using the currently active stream in `autograd/.../cudnn` (and assuming the backend only has 1 active stream per device, which is true for ArrayFire), this diff creates a new CUDA Stream and relatively synchronize it as necessary.

This way, we make no assumption on the streams, i.e., each input/output tensor could be on different stream, and the cudnn logic won't break.

When do we relativeSync?
1. Right before creating `DevicePtr`, sync cudnnstream against the tensor, but if reuse DevicePtr in the same function, no need to sync anymore, because it was synced at beginning, and will be up-to-date as cudnn use it.
2. For final output, sync its stream against cudnnstream, because the client of autograd expects the output `Tensor` to be ready to go.

Differential Revision: D37495449

fbshipit-source-id: d9bfec04d4bdf23604b6630793ad2d2ce6b3aac1