IR evaluator (#1022)
Summary:
We need to materialize the JIT graph when needed, e.g., (if client calls `tensor.toString()`).

Conceptually, `Evaluator`
1. does a **pass** through the JIT tree ("tree" since materialization is initiated from a single root node)
2. **dispatches** to another `TensorBackend`'s ops, e.g., `ScalarNode` gets dispatched to `backend.full(...)`
3. **annotate** node with tensor result (only if necessary, to save memory footprint).

The JIT graph is untouched, leaving room for
1. optimizations starting from other root node that involves a shared subtree
4. node graph printing with performance hotspots.

The Evaluator only depends on the JIT IR and the Tensor API -- it's agnostic to the JitTensor (IR parser).

Pull Request resolved: https://github.com/flashlight/flashlight/pull/1022

Test Plan:
Unit tests are added.

```
make JitEvaluatorTest JitTensorTest
./flashlight/fl/test/JitEvaluatorTest
./flashlight/fl/test/JitTensorTest
```

Reviewed By: bwasti, richjames0

Differential Revision: D40186292

fbshipit-source-id: d56576ec8b703fa432c292a47a33f794cf08e606