Implement onednn binop fusion pass (#1040)
Summary:
Leverage OneDNN's post-ops via backend-specific JIT optimization pass!

The pass itself is heavily documented, and the tests should also illustrate _what_ the pass does.

This PR focuses on a simple & extensible proof-of-concept -- there are a lot more we can do with this pass, as documented in the code.

### Test Plan (required)
Unit tests are added
```
make JitOneDnnOpFusionTest
```

Pull Request resolved: https://github.com/flashlight/flashlight/pull/1040

Reviewed By: bwasti

Differential Revision: D40768929

Pulled By: jacobkahn

fbshipit-source-id: 50f508cb1021b5cd73cbbd8c36a9ae9a41a66cd9